+++ userwidgets.ts
@@ -1,4 +1,4 @@
 export { User } from "./User"
 export { Application } from "./Application"
 export { Organization } from "./Organization"
-export { Client } from "./Client"
+export { ClientCollection } from "./ClientCollection"
